reset/requirements.txt


pymodbus
pyserial
